+++ CesiumVectorData/include/CesiumVectorData/VectorStyle.h
@@ -41,10 +41,11 @@ struct ColorStyle {
    * `ColorMode::Random`, this returns a randomized value obtained based on
    * the rules described in \ref ColorMode.
    *
-   * @param seed The seed for the random color to be generated, if `colorMode`
-   * is set to `Random`.
+   * @param randomColorSeed The seed for the random color to be generated, if
+   * `colorMode` is set to `Random`. The same color will always be returned for
+   * a given seed, but nearby seeds will not usually return nearby colors.
    */
-  CesiumUtility::Color getColor(uint32_t seed = 0) const;
+  CesiumUtility::Color getColor(size_t randomColorSeed = 0) const;
 };
 
 /** @brief The mode to use when interpreting a given line width. */