+++ CesiumVectorData/src/VectorStyle.cpp
@@ -3,23 +3,27 @@
 
 #include <cstdint>
 #include <optional>
-#include <random>
+#include <utility>
 
 using namespace CesiumUtility;
 
 namespace CesiumVectorData {
-Color ColorStyle::getColor(uint32_t seed) const {
+Color ColorStyle::getColor(size_t randomColorSeed) const {
   if (this->colorMode == ColorMode::Normal) {
     return this->color;
   }
 
-  std::mt19937 mt(seed);
-  std::uniform_real_distribution<float> dist(0.0f, 1.0f);
+  std::hash<size_t> hash{};
+  size_t h = hash(randomColorSeed);
+
+  float r = (h & 0xFF) / 255.0f;
+  float g = ((h >> 8) & 0xFF) / 255.0f;
+  float b = ((h >> 16) & 0xFF) / 255.0f;
 
   return Color{
-      (uint8_t)(dist(mt) * (float)this->color.r),
-      (uint8_t)(dist(mt) * (float)this->color.g),
-      (uint8_t)(dist(mt) * (float)this->color.b),
+      (uint8_t)(r * this->color.r),
+      (uint8_t)(g * this->color.g),
+      (uint8_t)(b * this->color.b),
       this->color.a};
 }
 VectorStyle::VectorStyle(const CesiumUtility::Color& color)