+++ CHANGES.md
@@ -8,6 +8,7 @@
 - `ViewUpdateResult` now holds pointers to const `Tile` instances.
 - The `slowlyGetCurrentStates` and `slowlyGetPreviousStates` methods of `TreeTraversalState` now return the state map with a raw pointer to a constant node as the key, even if the node pointer type is a smart pointer.
 - `DebugTileStateDatabase::recordTileState` now expects the states to be provided as `std::unordered_map<const Tile*, TileSelectionState>` instead of `std::unordered_map<IntrusivePointer<Tile>, TileSelectionState>`.
+- `VectorRasterizer::drawPolyline` now takes a `std::vector` instead of a `std::span`.
 
 ##### Additions :tada: